Update README (#75)
I find the instructions on here overall a bit confusing to follow...but is there a reason why we wanted people to be pointing to the legacy-stage consul? I updated that address to the mgmt-stage consul and wound up wiping out the entire thing. I think if you just leave that line out it will default to your local consul just fine...