fix(install): support arm64 for consul-template/vault
If architecture is arm64, install those binaries.

Use curl for downloading so we're just using one method for this script.
Keeping wget because other images may expect that to be available.

Cleanup unzip steps to only unzip the things we need.

Improve yum cleanup.

Remove docker cache comment as changing the content of the script should
accomplish the same thing.