feat(entrypoint): adds support for Vault auth from Lambda
When running a container Lambda, we can use the same AWS IAM auth
method. Lambda does not have a container credential variable, so check
for a different variable.

When I was testing, the Lambda IAM role name and the Vault role name was
difference, so it wasn't working as is. Add support for VAULT_ROLE if
you need to manually set the auth role

Add more information about installation and usage