feat: handle all signals
We can't guess all the signals that might be used, so forward everything
except for SIGCHILD, which is meant for the parent process

Add back in setting group pid. If it's directly connected, not having
this will cause double signals to be sent. Ignore shells and yarn
because they don't work well with this setting