refactor!: replace the current process instead of spawning a child process
We've ran into a lot of issues trying to have this forward signals and
reap child processes, everything that PID 1 needs to do. Instead of
having this contain everything we need, use dumb-init or tini or
something similar for PID 1 responsibilities.

Because we don't need to worry about all that, we can use exec to
replace the process with our child process instead of spawning a child
process, which we'd need to forward signals to

BREAKING CHANGE: won't reap child processes