feat(templates): Remove hardcoded env logic from consul export templates (legacy) (#241)
* chore(env): update prod export-vault template to work in any environment

There are currently three copies of the export templates in
{dev,prod,stage} subdirectories. The only difference between them is
dev/export-vault.ctmpl omits two export blocks. This adds a SERVICE_ENV
conditional check in the prod/export-vault.ctmpl so it can be used in
any environment. This will remove the need to have three copies in
env-specific directories.

* Remove duplicate templates and rename /prod/ to /templates/

* chore(env): Update template path references in scripts

After consolidating to one set of templates suitable for all
environments in the /templates/ directory, this updates the install.sh
and entrypoint.sh to remove env-specific logic for referencing
templates.

* fix(test): Remove non-working legacy dockerspec test